+++ src/helpers/convert-pcre.ts
@@ -8,8 +8,13 @@ export default function pcre(regex: string): RegExp {
 		finalRegex = finalRegex.replace(match, '');
 		[...flags].forEach(flag => finalFlags.add(flag));
 	}
-	finalRegex = finalRegex
-		.replace(/([*+]){2}/g, '$1') // ++ and *+ modifiers
-		.replace(/\\A/g, '^').replace(/\\Z/g, '$') // start- and end-of-file markers
+	// Remove invalid modifiers
+	finalRegex = finalRegex.replace(/([*+]){2}/g, '$1')
+	// Remove start/end-of-file markers
+	if (/\\[AZ]/.test(finalRegex)) {
+		finalRegex = finalRegex.replace(/\\A/g, '^').replace(/\\Z/g, '$');
+		finalFlags.delete('m');
+	}
+	else finalFlags.add('m');
 	return RegExp(finalRegex, [...finalFlags].join(''));
 }